eCloudWebStack/requirements.txt
asgiref==3.8.1
boto3==1.37.13
botocore==1.37.13
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.0
defusedxml==0.7.1
Django==5.1.7
django-allauth==65.8.0
django-storages==1.14.5
idna==3.10
jmespath==1.0.1
oauthlib==3.2.2
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.11.4
six==1.17.0
social-auth-core==4.6.1
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
